requirements.txt
azure-ai-vision-face==1.0.0b2
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.35.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
filelock==3.19.1
fonttools==4.59.2
idna==3.10
isodate==0.7.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.3.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.34.0
